penumbra_sdk_fee::component
Trait StateReadExt

pub trait StateReadExt: StateRead {
    // Provided methods
    fn get_fee_params<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<FeeParameters>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_gas_prices<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<GasPrices>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_alt_gas_prices<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<GasPrices>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn gas_prices_changed(&self) -> bool { ... }
    fn accumulated_base_fees_and_tips(&self) -> OrdMap<Id, (Amount, Amount)> { ... }
}

Available on crate feature component only.

This trait provides read access to fee-related parts of the Penumbra state store.

Provided Methods

fn get_fee_params<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<FeeParameters>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Gets the fee parameters from the JMT.

fn get_gas_prices<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<GasPrices>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Gets the current gas prices for the fee token.

fn get_alt_gas_prices<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Vec<GasPrices>>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Gets the current gas prices for alternative fee tokens.

fn gas_prices_changed(&self) -> bool

Returns true if the gas prices have been changed in this block.

fn accumulated_base_fees_and_tips(&self) -> OrdMap<Id, (Amount, Amount)>

The accumulated base fees and tips for this block, indexed by asset ID.

Object Safety

This trait is not object safe.